Removed `Source` and `RealmSource`. Moved sensor methods to `DatabaseHandler`
