+++ sense-android-library/src/nl/sense_os/datastorageengine/Sensor.java
@@ -15,7 +15,7 @@ public interface Sensor {
 
     String getUserId();
 
-    String getSourceId();
+    String getSource();
 
     SensorDataPoint.DataType getDataType();
